+++ CHANGELOG.md
@@ -65,6 +65,12 @@ There are following change types:
 
 - "Paste" action now supports special registers: `<C-w>` (word at cursor), `<C-a>` (WORD at cursor), `<C-l>` (line at cursor), `<C-f>` (filename at cursor).
 
+## mini.sessions
+
+### Refine
+
+- Update `read()` to write currently read session only if `MiniSessions.config.autowrite` is set to `true`.
+
 ## mini.surround
 
 ### Refine